+++ src/wowman/ui/cli.clj
@@ -37,14 +37,13 @@
   [_]
   ;; todo: move to core.clj
   (binding [http/*cache* (core/cache)]
-    (curseforge/download-all-addon-summaries (paths :curseforge-catalog-file))
-    (comment "disabled until the missing addons can be reconciled"
-             (let [output-file (paths :curseforge-catalog-file)
-                   catalog-data (curseforge-api/download-all-summaries-alphabetically)
-                   created (utils/datestamp-now-ymd)
-                   updated created
-                   formatted-catalog-data (catalog/format-catalog-data catalog-data created updated)]
-               (catalog/write-catalog-data output-file formatted-catalog-data)))))
+    ;;(curseforge/download-all-addon-summaries (paths :curseforge-catalog-file))
+    (let [output-file (paths :curseforge-catalog-file)
+          catalog-data (curseforge-api/download-all-summaries-alphabetically)
+          created (utils/datestamp-now-ymd)
+          updated created
+          formatted-catalog-data (catalog/format-catalog-data catalog-data created updated)]
+      (catalog/write-catalog-data output-file formatted-catalog-data))))
 
 (defmethod action :update-curseforge-catalog
   [_]