+++ test.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+# (always ratchet threshold upwards)
+lein cloverage --fail-threshold 72 --html