+++ test/wowman/wowinterface_test.clj
@@ -46,13 +46,19 @@
           expected []
           num-addons 25
           first-addon {:uri "https://www.wowinterface.com/downloads/info25079",
-                       :name "rotation-master", :label "Rotation Master",
+                       :name "rotation-master",
+                       :label "Rotation Master",
                        :updated-date "2019-07-29T21:37:00Z",
-                       :download-count 80, :category-list #{"dummy"}}
+                       :download-count 80,
+                       :category-list #{"dummy"}
+                       :source-id 25079}
           last-addon  {:uri "https://www.wowinterface.com/downloads/info24805",
-                       :name "mattbars-mattui", :label "MattBars (MattUI)",
+                       :name "mattbars-mattui",
+                       :label "MattBars (MattUI)",
                        :updated-date "2018-10-30T17:56:00Z",
-                       :download-count 1911, :category-list #{"dummy"}}]
+                       :download-count 1911,
+                       :category-list #{"dummy"}
+                       :source-id 24805}]
       (with-fake-routes-in-isolation fake-routes
         (let [results (wowinterface/scrape-addon-page category page)]
           (is (= num-addons (count results)))
@@ -74,6 +80,6 @@
                            :version "8.2.10",
                            :interface-version 80200})]
       (with-fake-routes-in-isolation fake-routes
-        ;; there is a suspicious 'pause' here with the http/download take 461 msecs of 480 msec execution
+        ;; the suspicious pause here taking 461 msecs of 480 msec execution is the conversion of the body into a html snippet
         (let [results (wowinterface/expand-summary addon-summary)]
           (is (= expected results)))))))